feat(network): transaction sync restructuring
