+++ network/src/transactions/mod.rs
@@ -14,5 +14,5 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkOS library. If not, see <https://www.gnu.org/licenses/>.
 
-pub mod blocks;
-pub use blocks::*;
+pub mod transactions;
+pub use transactions::*;
